music_app/requirements.txt
amqp==2.5.2
api==0.0.7
app==0.0.1
appdirs==1.4.3
attrs==19.3.0
billiard==3.6.1.0
black==19.10b0
bleach==3.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coverage==4.5.4
Django==2.2.7
django-celery-results==1.1.2
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
docutils==0.15.2
filelock==3.0.12
idna==2.8
importlib-metadata==0.23
kombu==4.6.6
more-itertools==7.2.0
nose==1.3.7
packaging==19.2
pathspec==0.6.0
pkginfo==1.5.0.1
pluggy==0.13.0
psycopg2==2.8.4
py==1.8.0
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.5
pytz==2019.3
readme-renderer==24.0
redis==3.3.11
regex==2019.11.1
requests==2.22.0
requests-toolbelt==0.9.1
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
tox==3.14.0
tqdm==4.38.0
twine==2.0.0
typed-ast==1.4.0
urllib3==1.25.7
views==0.3
vine==1.3.0
virtualenv==16.7.7
webencodings==0.5.1
zipp==0.6.0
